NN-2323: Remove the concept of CLIENT and move relevant components to main app
