+++ Dockerfile
@@ -1,12 +1,10 @@
 FROM node:10.15-slim
 LABEL maintainer="HMPPS Digital Studio <info@digital.justice.gov.uk>"
 
-ARG CLIENT
 ARG BUILD_NUMBER
 ARG GIT_REF
 ARG NODE_ENV
 
-ENV CLIENT ${CLIENT:-hmpps}
 ENV BUILD_NUMBER ${BUILD_NUMBER:-1_0_0}
 ENV GIT_REF ${GIT_REF:-dummy}
 
@@ -19,7 +17,6 @@ WORKDIR /app
 ADD . .
 
 RUN yarn --frozen-lockfile && \
-    export CLIENT=${CLIENT} && \
     yarn build && \
     export BUILD_NUMBER=${BUILD_NUMBER} && \
     export GIT_REF=${GIT_REF} && \
