--- app/customers/syscon/_font_stack.scss
@@ -1,46 +1,0 @@
-//  GOV.UK font stacks, referred to in typography.scss
-
-// New Transport Light
-$nta-light: Inter UI, Arial, sans-serif;
-$nta-light-tabular: "ntatabularnumbers", $nta-light;
-
-// Helvetica Regular
-$helvetica-regular: "HelveticaNeue", "Helvetica Neue", "Arial", "Helvetica", sans-serif;
-
-// Allow font stack to be overridden
-// Not all apps using toolkit use New Transport
-$toolkit-font-stack: $nta-light !default;
-$toolkit-font-stack-tabular: $nta-light-tabular !default;
-
-// Font reset for print
-$print-reset: sans-serif;
-
-// Fallback variable names after renaming previous uppercase names to be lowercase
-// @deprecated, please only use the lowercase versions
-// Make an exception to the linting as these are still used a lot
-// scss-lint:disable NameFormat
-$NTA-Light: $nta-light;
-$NTA-Light-Tabular: $nta-light-tabular;
-$Helvetica-Regular: $helvetica-regular;
-$Print-reset: $print-reset;
-
-
-@font-face {
-  font-family: Inter UI;
-  src: url('/fonts/Inter-UI/Inter-UI-Regular.woff') format('woff');
-  font-weight: normal;
-  font-style: normal;
-}
-@font-face {
-  font-family: Inter UI;
-  src: url('/fonts/Inter-UI/Inter-UI-Bold.woff') format('woff');
-  font-weight: bold;
-}
-
-
-body{
-  font-family: Inter UI, com Arial, sans-serif;
-  margin: 0;
-  height: 100%;
-}
-
