--- app/customers/syscon/app/components/Header/index.scss
@@ -1,79 +1,0 @@
-.banner {
-  width: 100%;
-  background: rgb(0, 39, 72);
-  color: white;
-  z-index: 999;
-
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-
-  @include media-breakpoint-down(md) {
-    justify-content: start;
-  }
-
-  @include media-breakpoint-down(xs) {
-    .desktop-only {
-      display: none;
-    }
-  }
-
-  .brand-header {
-    display: flex;
-    flex-direction: row;
-
-    .section {
-      margin-left: 15px;
-      margin-right: 15px;
-
-      .divider {
-        border-right: solid 2px white;
-        margin-top: 13px;
-        height: 60%;
-        display: block;
-      }
-    }
-
-    h1 {
-      font-size: 24px;
-      font-weight: bold;
-    }
-
-    img {
-      width: 40px;
-      height: 40px;
-      margin-top: 13px;
-    }
-
-    @include media-breakpoint-down(md) {
-      h1 {
-        font-size: 19px;
-      }
-      img {
-        margin-top: 6px;
-      }
-    }
-  }
-
-  .desktop-menu {
-    height: 60px;
-    display: block;
-  }
-
-  .mobile-menu {
-    display: none;
-  }
-
-  @include media-breakpoint-down(md) {
-    .desktop-menu {
-      display: none;
-    }
-    .mobile-menu {
-      position: absolute;
-      top: 0;
-      right: 0;
-      display: block;
-    }
-  }
-}
