--- app/customers/syscon/app/product-globals.js
@@ -1,4 +1,0 @@
-export default {
-  customerName: 'SYSCON',
-  serviceName: 'Prison Manager',
-}
