--- app/customers/syscon/colours/_organisation.scss
@@ -1,104 +1,0 @@
-// We use `websafe` to mean strong enough contrast against white to
-// be used for copy and meet the AAA (large text) and AA (smaller
-// copy) WCAG guidelines.
-
-$attorney-generals-office: #9f1888;
-$attorney-generals-office-websafe: #a03a88;
-$cabinet-office: #005abb;
-$cabinet-office-websafe: #347da4;
-$civil-service: #af292e;
-$department-for-business-innovation-skills: #003479;
-$department-for-business-innovation-skills-websafe: #347da4;
-$department-for-communities-and-local-government: #00857e;
-$department-for-communities-and-local-government-websafe: #37836e;
-$department-for-culture-media-sport: #d40072;
-$department-for-culture-media-sport-websafe: #a03155;
-$department-for-education: #003a69;
-$department-for-education-websafe: #347ca9;
-$department-for-environment-food-rural-affairs: #00a33b;
-$department-for-environment-food-rural-affairs-websafe: #008938;
-$department-for-international-development: #002878;
-$department-for-international-development-websafe: #405e9a;
-$department-for-international-trade: #CF102D;
-$department-for-international-trade-websafe: $link-colour;
-$department-for-transport: #006c56;
-$department-for-transport-websafe: #398373;
-$department-for-work-pensions: #00beb7;
-$department-for-work-pensions-websafe: #37807b;
-$department-of-energy-climate-change: #009ddb;
-$department-of-energy-climate-change-websafe: #2b7cac;
-$department-of-health: #00ad93;
-$department-of-health-websafe: #39836e;
-$foreign-commonwealth-office: #003e74;
-$foreign-commonwealth-office-websafe: #406e97;
-$government-equalities-office: #9325b2;
-$hm-government: #0076c0;
-$hm-government-websafe: #347da4;
-$hm-revenue-customs: #009390;
-$hm-revenue-customs-websafe: #008670;
-$hm-treasury: #af292e;
-$hm-treasury-websafe: #832322;
-$home-office: #9325b2;
-$home-office-websafe: #9440b2;
-$ministry-of-defence: #4d2942;
-$ministry-of-defence-websafe: #5a5c92;
-$ministry-of-justice: #231f20;
-$ministry-of-justice-websafe: #5a5c92;
-$northern-ireland-office: #002663;
-$northern-ireland-office-websafe: #3e598c;
-$office-of-the-advocate-general-for-scotland: #002663;
-$office-of-the-advocate-general-for-scotland-websafe: $link-colour;
-$office-of-the-leader-of-the-house-of-lords: #9c132e;
-$office-of-the-leader-of-the-house-of-lords-websafe: #c2395d;
-$scotland-office: #002663;
-$scotland-office-websafe: #405c8a;
-// Note: the 'the' part here will get dropped
-$the-office-of-the-leader-of-the-house-of-commons: #317023;
-$the-office-of-the-leader-of-the-house-of-commons-websafe: #005f8f;
-$uk-export-finance: #005747;
-$uk-export-finance-websafe: $link-colour;
-$uk-trade-investment: #C80651;
-$uk-trade-investment-websafe: $link-colour;
-$wales-office: #a33038;
-$wales-office-websafe: #7a242a;
-
-// All organisation colours in a list
-// (class_name, brand colour, WCAG acceptible text colour)
-//
-// example usage:
-// @each $organisation in $all-organisation-brand-colours {
-//    .#{nth($organisation, 1)} {
-//      border-color: nth($organisation, 2);
-//    }
-// }
-
-$all-organisation-brand-colours:
-  'attorney-generals-office' $attorney-generals-office $attorney-generals-office-websafe,
-  'cabinet-office' $cabinet-office $cabinet-office-websafe,
-  'civil-service' $civil-service $civil-service,
-  'department-for-business-innovation-skills' $department-for-business-innovation-skills $department-for-business-innovation-skills-websafe,
-  'department-for-communities-and-local-government' $department-for-communities-and-local-government $department-for-communities-and-local-government-websafe,
-  'department-for-culture-media-sport' $department-for-culture-media-sport $department-for-culture-media-sport-websafe,
-  'department-for-education' $department-for-education $department-for-education-websafe,
-  'department-for-environment-food-rural-affairs' $department-for-environment-food-rural-affairs $department-for-environment-food-rural-affairs-websafe,
-  'department-for-international-development' $department-for-international-development $department-for-international-development-websafe,
-  'department-for-international-trade' $department-for-international-trade $department-for-international-trade-websafe,
-  'department-for-transport' $department-for-transport $department-for-transport-websafe,
-  'department-for-work-pensions' $department-for-work-pensions $department-for-work-pensions-websafe,
-  'department-of-energy-climate-change' $department-of-energy-climate-change $department-of-energy-climate-change-websafe,
-  'department-of-health' $department-of-health $department-of-health-websafe,
-  'foreign-commonwealth-office' $foreign-commonwealth-office $foreign-commonwealth-office-websafe,
-  'hm-government' $hm-government $hm-government-websafe,
-  'hm-revenue-customs' $hm-revenue-customs $hm-revenue-customs-websafe,
-  'hm-treasury' $hm-treasury $hm-treasury-websafe,
-  'home-office' $home-office $home-office-websafe,
-  'ministry-of-defence' $ministry-of-defence $ministry-of-defence-websafe,
-  'ministry-of-justice' $ministry-of-justice $ministry-of-justice-websafe,
-  'northern-ireland-office' $northern-ireland-office $northern-ireland-office-websafe,
-  'office-of-the-advocate-general-for-scotland' $office-of-the-advocate-general-for-scotland $office-of-the-advocate-general-for-scotland-websafe,
-  'office-of-the-leader-of-the-house-of-lords' $office-of-the-leader-of-the-house-of-lords $office-of-the-leader-of-the-house-of-lords-websafe,
-  'scotland-office' $scotland-office $scotland-office-websafe,
-  'the-office-of-the-leader-of-the-house-of-commons' $the-office-of-the-leader-of-the-house-of-commons $the-office-of-the-leader-of-the-house-of-commons-websafe,
-  'uk-export-finance' $uk-export-finance $uk-export-finance-websafe,
-  'uk-trade-investment' $uk-trade-investment $uk-trade-investment-websafe,
-  'wales-office' $wales-office $wales-office-websafe;
