--- app/customers/syscon/index.scss
@@ -1,3 +1,0 @@
-@import 'colours/palette';
-@import 'colours/organisation';
-@import './_font_stack';
