+++ app/product-globals.js
