+++ package.json
@@ -14,16 +14,13 @@
     "postinstall": "npm run build:dll",
     "build": "cross-env NODE_ENV=production webpack --config internals/webpack/webpack.prod.babel.js --color -p",
     "build:analyse": "BUNDLE_ANALYSE=true npm run build:hmpps",
-    "build:syscon": "cross-env CLIENT=syscon NODE_ENV=production webpack --config internals/webpack/webpack.prod.babel.js --color -p",
-    "build:hmpps": "cross-env CLIENT=hmpps NODE_ENV=production webpack --config internals/webpack/webpack.prod.babel.js --color -p",
     "build:dll": "node ./internals/scripts/dependencies.js",
-    "start": "cross-env CLIENT=hmpps NODE_ENV=development node server",
-    "start-feature": "cross-env CLIENT=hmpps NODE_ENV=development PORT=3007 USE_OF_FORCE_URL=http://use-of-force PATHFINDER_URL=http://pathfinder USE_OF_FORCE_PRISONS=LEI OAUTH_ENDPOINT_URL=http://localhost:19090/auth/ API_ENDPOINT_URL=http://localhost:18080/ KEYWORKER_API_URL=http://localhost:18081/ PRISON_STAFF_HUB_UI_URL=http://localhost:18082/ CASENOTES_API_URL=http://localhost:18083 ALLOCATION_MANAGER_ENDPOINT_URL=http://localhost:18084/ node server | bunyan -o short",
+    "start": "cross-env NODE_ENV=development node server",
+    "start-feature": "cross-env NODE_ENV=development PORT=3007 USE_OF_FORCE_URL=http://use-of-force PATHFINDER_URL=http://pathfinder USE_OF_FORCE_PRISONS=LEI OAUTH_ENDPOINT_URL=http://localhost:19090/auth/ API_ENDPOINT_URL=http://localhost:18080/ KEYWORKER_API_URL=http://localhost:18081/ PRISON_STAFF_HUB_UI_URL=http://localhost:18082/ CASENOTES_API_URL=http://localhost:18083 ALLOCATION_MANAGER_ENDPOINT_URL=http://localhost:18084/ node server | bunyan -o short",
     "start:tunnel": "cross-env NODE_ENV=development ENABLE_TUNNEL=true node server",
     "start:production": "npm run test && npm run build && npm run start:prod",
     "start:prod": "cross-env NODE_ENV=production node server",
     "presetup": "npm i chalk shelljs",
-    "setup": "node ./internals/scripts/setup.js",
     "postsetup": "npm run build:dll",
     "clean": "shjs ./internals/scripts/clean.js",
     "clean:all": "npm run analyze:clean && npm run test:clean && npm run build:clean",
